requirements.txt
openai
python-dotenv
fastapi
uvicorn
youtube-transcript-api